<NLBackend> #43 refactor: change check delete tag is valid place

+++ NestLedger/src/main/java/org/xxooooxx/nestledger/service/tag/implementation/TagServiceImpl.java
@@ -80,6 +80,9 @@ public class TagServiceImpl implements TagService {
         if (tagDB == null) {
             throw new CustomException(CustomExceptionEnum.TAG_NOT_FOUND);
         }
+        if (tagDB.getUsingCount() > 0) {
+            throw new CustomException(CustomExceptionEnum.TAG_STILL_IN_USE);
+        }
         checkOperationValid(tagDB.getLedgerId());
         tagDao.deleteTag(tagId);
     }
